Fix spelling of interrupt
